add filter components
