+++ src/pages/shop/sections/filter/CategoryFilter.js
@@ -1,0 +1,7 @@
+const CategoryFilter = () => (
+    <div>
+        Categories
+    </div>
+)
+
+export default CategoryFilter
