+++ src/pages/shop/sections/filter/ColorFilter.js
@@ -1,0 +1,7 @@
+const ColorFilter = () => (
+    <div>
+        Color
+    </div>
+)
+
+export default ColorFilter
