+++ src/pages/shop/sections/filter/PriceFilter.js
@@ -1,0 +1,7 @@
+const PriceFilter = () => (
+    <div>
+        Price
+    </div>
+)
+
+export default PriceFilter
