+++ src/pages/shop/sections/filter/SearchFilter.js
@@ -1,0 +1,7 @@
+const SearchFilter = () => (
+    <div>
+        Search
+    </div>
+)
+
+export default SearchFilter
