+++ src/pages/shop/sections/filter/SizeFilter.js
@@ -1,0 +1,7 @@
+const SizeFilter = () => (
+    <div>
+        Size
+    </div>
+)
+
+export default SizeFilter
